Move parameter tuning into separate class
